analisis/cambios propuestas/DAMICH/damich-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
sqlalchemy==2.0.36
fpdf2==2.8.1
